requirements.txt
streamlit
openai
databricks-sql-connector[pyarrow]
pandas
google-generativeai
tabulate
duckdb
